data-managment/requirements.txt
delta-spark==3.2.0
pyspark==3.5.1
